requirements.txt
antlr4-python3-runtime==4.9.3
argcomplete==3.2.2
asttokens==2.4.1
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cma==3.3.0
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
coverage==7.4.1
cycler==0.12.1
dask==2024.1.0
debugpy==1.8.0
decorator==5.1.1
dill==0.3.8
distlib==0.3.8
distributed==2024.1.0
docopt==0.6.2
executing==2.0.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
identify==2.5.33
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.0
kiwisolver==1.4.5
leap-ec==0.8.1
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
multiprocess==0.70.15
nest-asyncio==1.5.8
networkx==3.2.1
nodeenv==1.8.0
numpy==1.26.3
omegaconf==2.3.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
partd==1.4.1
pillow==10.2.0
pipx==1.4.3
platformdirs==4.1.0
pluggy==1.4.0
pre-commit==3.6.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pyade-python==1.1
Pygments==2.17.2
pyparsing==3.1.1
pytest==8.0.0
pytest-cov==4.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rich==13.7.0
scipy==1.12.0
six==1.16.0
smac==2.0.2
sortedcontainers==2.4.0
stack-data==0.6.3
structlog==24.1.0
sympy==1.12
tblib==3.0.0
toolz==0.12.0
tornado==6.4
traitlets==5.14.1
treelib==1.7.0
tzdata==2023.4
urllib3==2.1.0
userpath==1.9.1
virtualenv==20.25.0
wcwidth==0.2.12
zict==3.0.0
zipp==3.17.0
